requirements.txt
groq
python-docx
PyPDF2